Refactor to extract viewportAggregation calculations to its own module
